// frontend/src/types.ts
export interface Task {
  id: number;
  title: string;
  description: string;
  completed: boolean;
  created_at: string;
  updated_at: string;
}

export interface TaskList {
  count: number;
  next: string | null;
  previous: string | null;
  results: Task[];
}

export interface TokenPair {
  access: string;
  refresh: string;
}